requirements.txt
Django==1.4
MySQL-python==1.2.3
certifi==0.0.8
chardet==1.0.1
distribute==0.6.19
django-reroute==1.1.1
ipdb==0.6.1
ipython==0.12.1
psycopg2==2.4.5
requests==0.11.2
wsgiref==0.1.2